+++ CHANGELOG.md
@@ -5,5 +5,6 @@
 - **Breaking:** removed `robots.Bot`
 - **Breaking:** Node 20+ is now required
 - **Breaking:** require ECMAScript modules (no more CommonJS)
+- **Breaking:** middleware only responds to `GET` requests to `/robots.txt`
 
 This changelog was created after the release of 0.3.0.